requirements.txt
dotenv==0.9.9
google-cloud-storage==2.19.0
python-dotenv==1.2.1
requests==2.32.5

